@@ -1,16 +1,18 @@
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
-    id("org.springframework.boot") version "2.4.3"
-    id("io.spring.dependency-management") version "1.0.11.RELEASE"
-    kotlin("jvm") version "1.4.30"
-    kotlin("plugin.spring") version "1.4.30"
-    maven
+    id("org.springframework.boot") version "3.1.4"
+    id("io.spring.dependency-management") version "1.1.3"
+    kotlin("jvm") version "1.8.22"
+    kotlin("plugin.spring") version "1.8.22"
+    //maven
 }
 
 group = "com.vmware.tanzu.data"
 version = "0.0.1-SNAPSHOT"
-java.sourceCompatibility = JavaVersion.VERSION_11
+java {
+    sourceCompatibility = JavaVersion.VERSION_17
+}
 
 repositories {
     maven { url = uri("https://oss.sonatype.org/content/repositories/snapshots") }
@@ -19,19 +21,20 @@ repositories {
     mavenLocal()
 }
 
-extra["springCloudVersion"] = "2020.0.1"
+extra["springCloudVersion"] = "2021.0.0"
 
 
 
 dependencies {
 //    implementation("org.springdoc:springdoc-openapi-ui:1.5.2")
-    implementation("com.github.nyla-solutions:nyla.solutions.core:1.4.3")
+    implementation("org.springdoc:springdoc-openapi:2.2.0")
+    implementation("com.github.nyla-solutions:nyla.solutions.core:1.4.4")
     implementation(project(":components:IoT-connected-vehicles-domains"))
     implementation(project(":components:vehicle-messaging"))
     implementation(project(":components:messaging-streaming"))
     implementation(project(":components:vehicle-messaging-streaming"))
 //    implementation("org.springframework.boot:spring-boot-starter-amqp")
-    implementation("com.rabbitmq:stream-client:0.1.0-SNAPSHOT")
+    implementation("com.rabbitmq:stream-client:0.3.0")
     implementation("org.apache.qpid:proton-j:0.33.8")
     implementation("org.xerial.snappy:snappy-java:1.1.8.4")
     implementation("org.lz4:lz4-java:1.8.0")
@@ -53,7 +56,7 @@ dependencyManagement {
 tasks.withType<KotlinCompile> {
     kotlinOptions {
         freeCompilerArgs = listOf("-Xjsr305=strict")
-        jvmTarget = "11"
+        jvmTarget = "17"
     }
 }
 
