@@ -1,17 +1,20 @@
 plugins {
-    kotlin("jvm") version "1.4.31"
+    kotlin("jvm") version "1.8.22"
+    kotlin("plugin.spring") version "1.8.22"
 }
 
 group = "com.vmware.tanzu.data"
 version = "0.0.1-SNAPSHOT"
-java.sourceCompatibility = JavaVersion.VERSION_11
+java {
+    sourceCompatibility = JavaVersion.VERSION_17
+}
 
 repositories {
     mavenCentral()
 }
 dependencies {
-    implementation("org.apache.geode:geode-core:1.13.1");
-    implementation("org.apache.geode:geode-lucene:1.13.1")
+    implementation("org.apache.geode:geode-core:1.13.7");
+    implementation("org.apache.geode:geode-lucene:1.13.7")
     implementation("org.junit.jupiter:junit-jupiter:5.4.2");
     implementation("org.apache.logging.log4j:log4j-api:2.13.1");
     testImplementation("org.mockito:mockito-junit-jupiter:3.8.0");
