@@ -1,28 +1,31 @@
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
-    kotlin("jvm") version "1.4.30"
+//    id("org.springframework.boot") version "3.1.4"
+    id("io.spring.dependency-management") version "1.1.3"
+    kotlin("jvm") version "1.8.22"
+//    kotlin("plugin.spring") version "1.8.22"
 }
 
 group = "com.vmware.tanzu.data"
 version = "0.0.1-SNAPSHOT"
-java.sourceCompatibility = JavaVersion.VERSION_11
+java {
+    sourceCompatibility = JavaVersion.VERSION_17
+}
 
 repositories {
-    maven { url = uri("https://oss.sonatype.org/content/repositories/snapshots") }
-    maven { url = uri("https://packagecloud.io/rabbitmq/maven-milestones/maven2") }
     mavenCentral()
     mavenLocal()
 
 }
 
-extra["springCloudVersion"] = "2020.0.1"
+extra["springCloudVersion"] = "2021.0.0"
 
 dependencies {
-    implementation("org.springdoc:springdoc-openapi-ui:1.5.2")
-    implementation("com.github.nyla-solutions:nyla.solutions.core:1.4.3")
-    implementation(project(":components:IoT-connected-vehicles-domains"))
+    implementation("org.springframework.boot:spring-boot-starter:3.1.4")
     implementation("org.jetbrains.kotlin:kotlin-reflect")
+    implementation("com.github.nyla-solutions:nyla.solutions.core:1.4.4")
+    implementation(project(":components:IoT-connected-vehicles-domains"))
     implementation("org.jetbrains.kotlin:kotlin-stdlib-jdk8")
     implementation(project(":components:vehicle-messaging"))
     implementation("org.eclipse.paho:org.eclipse.paho.client.mqttv3:1.2.0")
@@ -35,7 +38,7 @@ dependencies {
 tasks.withType<KotlinCompile> {
     kotlinOptions {
         freeCompilerArgs = listOf("-Xjsr305=strict")
-        jvmTarget = "11"
+        jvmTarget = "17"
     }
 }
 
