@@ -3,7 +3,8 @@ package com.vmware.tanzu.data.IoT.vehicles.source
 import com.vmware.tanzu.data.IoT.vehicles.domains.Vehicle
 import com.vmware.tanzu.data.IoT.vehicles.messaging.telemetry.repositories.VehicleTelemetryRepository
 import org.hibernate.cfg.AvailableSettings
-import org.hibernate.dialect.PostgreSQL9Dialect
+//import org.hibernate.dialect.PostgreSQL9Dialect
+//import org.hibernate.dialect.PostgreSQL9Dialect
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
@@ -28,7 +29,7 @@ class JdbcJpaConfig {
         )
         //these needed to be added to have all hibernate config done in one place.
         em.jpaPropertyMap[AvailableSettings.CURRENT_SESSION_CONTEXT_CLASS] = SpringSessionContext::class.java.name
-        em.jpaPropertyMap[AvailableSettings.DIALECT] = PostgreSQL9Dialect::class.java.name
+//        em.jpaPropertyMap[AvailableSettings.DIALECT] = PostgreSQL9Dialect::class.java.name
         val vendor = HibernateJpaVendorAdapter()
         vendor.setShowSql(true)
         em.jpaVendorAdapter = vendor
