@@ -3,8 +3,8 @@ package com.vmware.tanzu.data.IoT.vehicles.messaging.streaming.routing
 import com.rabbitmq.stream.*
 import com.vmware.tanzu.data.IoT.vehicles.messaging.streaming.creational.StreamSetup
 import nyla.solutions.core.patterns.creational.Creator
-import org.springframework.beans.factory.annotation.Value
-import org.springframework.stereotype.Component
+//import org.springframework.beans.factory.annotation.Value
+//import org.springframework.stereotype.Component
 import java.time.Duration
 import java.util.function.Function
 
@@ -14,30 +14,30 @@ import java.util.function.Function
  *
  * @author Gregory Green
  */
-@Component
+//@Component
 class RabbitStreamingRoutingFunctionHandler(
-    @Value("\${rabbitmq.streaming.routing.input.stream.name}")
+//    @Value("\${rabbitmq.streaming.routing.input.stream.name}")
     private val inputStreamName: String,
     private val inputStreamSetup: StreamSetup,
     private val inputEnvCreator: Creator<Environment>,
-    @Value("\${rabbitmq.streaming.routing.output.stream.name}")
+//    @Value("\${rabbitmq.streaming.routing.output.stream.name}")
     private val outputStreamName: String,
     private val outputEnvCreator: Creator<Environment>,
     private val outputStreamSetup: StreamSetup,
     private val routingFunction: Function<ByteArray, ByteArray>,
-    @Value("\${spring.application.name}")
+//    @Value("\${spring.application.name}")
     private val applicationName: String,
-    @Value("\${rabbitmq.streaming.routing.input.replay.offset}")
+//    @Value("\${rabbitmq.streaming.routing.input.replay.offset}")
     private val offset: Long = 0,
-    @Value("\${rabbitmq.streaming.routing.input.replay.bool}")
+//    @Value("\${rabbitmq.streaming.routing.input.replay.bool}")
     private val replay: Boolean = false,
     private val handler: ConfirmationHandler = ConfirmationHandler { status ->
         if (!status.isConfirmed)
             println("ERROR: $status.code NOT confirmed}")
     },
-    @Value("\${rabbitmq.streaming.routing.input.messageCountBeforeStorage:50000}")
+//    @Value("\${rabbitmq.streaming.routing.input.messageCountBeforeStorage:50000}")
     private var messageCountBeforeStorage: Int = 50_000,
-    @Value("\${rabbitmq.streaming.routing.input.flushIntervalDurationSecs:10}")
+//    @Value("\${rabbitmq.streaming.routing.input.flushIntervalDurationSecs:10}")
     private var flushIntervalDurationSecs: Long = 10
 ) : MessageHandler, AutoCloseable {
 
