@@ -2,9 +2,9 @@ package com.vmware.tanzu.data.IoT.vehicles.generator
 
 import com.vmware.tanzu.data.IoT.vehicles.domains.Vehicle
 import com.vmware.tanzu.data.IoT.vehicles.messaging.vehicle.publisher.VehicleSender
-import org.springframework.stereotype.Service
+//import org.springframework.stereotype.Service
 
-@Service
+//@Service
 interface  VehicleRider {
 
     /**
